requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==5.0.1
django-cors-headers==4.3.1
django-json-widget==1.1.1
django-material-admin==1.8.6
djangorestframework==3.14.0
drf-yasg==1.21.7
future==0.18.3
gevent==23.9.1
greenlet==3.0.3
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
material==0.1
packaging==23.2
pillow==10.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
rpds-py==0.17.1
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.9.0
uritemplate==4.1.1
whitenoise==6.6.0
zope.event==5.0
zope.interface==6.1
